requirements.txt
fastapi
uvicorn
gunicorn
supabase
requests
fastembed
python-dotenv
pydantic